Refactor nox_xxx_mobActionDodge_544640.

+++ src/legacy/object.go
@@ -345,7 +345,7 @@ func Nox_xxx_unitPostCreateNotify_4E7F10(a1 *server.Object) {
 	nox_xxx_unitPostCreateNotify_4E7F10(asObjectC(a1))
 }
 func Nox_xxx_buffApplyTo_4FF380(a1 *server.Object, a2 server.EnchantID, a3 int, a4 int) {
-	nox_xxx_buffApplyTo_4FF380(asObjectC(a1), int32(a2), int16(a3), int8(a4))
+	nox_xxx_buffApplyTo_4FF380(asObjectC(a1), a2, int16(a3), int8(a4))
 }
 func Nox_xxx_spellBuffOff_4FF5B0(a1 *server.Object, a2 server.EnchantID) {
 	nox_xxx_spellBuffOff_4FF5B0(asObjectC(a1), int32(a2))
